+++ desktop/src/main/java/com/onecat/burst/ui/displays/RegionInfluencerDisplay.java
@@ -9,6 +9,7 @@ import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.utils.Align;
 import com.badlogic.gdx.utils.Array;
 import com.onecat.burst.Main;
+import com.onecat.burst.ui.widgets.ListPicker;
 import com.onecat.burst.world.ParticleManager;
 import java.util.ArrayList;
 import java.util.Objects;