+++ desktop/src/main/java/com/onecat/burst/ui/displays/RegularEmitterDisplay.java
@@ -6,6 +6,7 @@ import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.ui.CheckBox;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
+import com.onecat.burst.ui.widgets.DisplayComponent;
 
 public class RegularEmitterDisplay extends ControllerComponentDisplay {