+++ desktop/src/main/java/com/onecat/burst/ui/displays/SpawnInfluencerDisplay.java
@@ -7,6 +7,7 @@ import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.utils.Align;
+import com.onecat.burst.ui.widgets.DisplayComponent;
 
 public class SpawnInfluencerDisplay extends ControllerComponentDisplay {