+++ desktop/src/main/java/com/onecat/burst/ui/widgets/ListPicker.java
@@ -1,4 +1,4 @@
-package com.onecat.burst.ui.displays;
+package com.onecat.burst.ui.widgets;
 
 import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.Stage;